+++ src/App.js
@@ -1,15 +1,39 @@
 import React from 'react';
-import TodoList from './components/TodoComponents/TodoList';
+//import TodoList from './components/TodoComponents/TodoList';
+import Todo from "./components/TodoComponents/Todo";
+import TodoForm from "./components/TodoComponents/TodoForm";
+
 
 class App extends React.Component {
-  // you will need a place to store your state in this component.
-  // design `App` to be the parent component of your application.
-  // this component is going to take care of state, and any change handlers you need to work with your state
+  constructor(props) {
+    super(props);
+    this.state = {
+      completed: false,
+      inputText: "",
+      id:Date.now()
+    }
+  }
+  addToInput(input) {
+    this.setState({ input:  + input });
+};
+
+
   render() {
     return (
       <div>
         <h2>Welcome to your Todo App!</h2>
-        <TodoList tasks={TASKS}/>
+        <div>
+          <Todo 
+            tasks={this.props.tasks}
+            completed={this.state.completed}
+           />
+          <TodoForm 
+          completed={this.state.completed}
+          inputText={this.state.input}
+          onClick={() => this.addToInput(inputText)}
+          
+          />
+        </div>
       </div>
     );
   }
@@ -17,17 +41,6 @@ class App extends React.Component {
 
 
 
-const TASKS = [
-  {
-    task: 'Organize Garage',
-    id: 1528817077286,
-    completed: false
-  },
-  {
-    task: 'Bake Cookies',
-    id: 1528817084358,
-    completed: true
-  }
-];
+
 
 export default App;
