+++ src/components/TodoComponents/TodoForm.js
@@ -1,11 +1,22 @@
 import React, { Component } from 'react';
 
+
 class TodoForm extends React.Component {
     render() {
+        const inputText = this.props.inputText;
+        
         return (
             <form>
-                <input type="text" placeholder="Add ToDo"></input>
-                <input type="button" value="Add Tasks"/>
+                <input 
+                value={inputText}
+                type="text" 
+                placeholder="Add ToDo"
+                />
+                <input 
+                type="button" 
+                //value="Add Tasks"
+                onClick = {this.value.onClick}
+                />
                 <input type="button" value="Remove Completed"/>
             </form>
         )
