+++ src/components/TodoComponents/TodoList.js
@@ -1,19 +1,1 @@
-import React, { Component } from 'react';
-import Todo from "./Todo";
-import TodoForm from "./TodoForm";
-
-
-
-class TodoList extends React.Component {
-    render() {
-        return (
-            <div>
-                <Todo tasks={this.props.tasks}/>
-                <TodoForm />
-            </div>
-        )
-    }
-}
-
-export default TodoList;
-
+//not using this file, Todo is a single class with a function for task list element
