+++ src/index.js
@@ -2,5 +2,17 @@ import React from 'react';
 import { render } from 'react-dom';
 import App from './App';
 
-render(<App />, document.getElementById('root'));
+const TASKS = [
+    {
+      task: 'Organize Garage',
+      id: 1528817077286,
+      completed: false
+    },
+    {
+      task: 'Bake Cookies',
+      id: 1528817084358,
+      completed: true
+    }
+  ];
+render(<App tasks={TASKS} />, document.getElementById('root'));
